feat(vulnerability): Unify PC/mobile views for vulnerability management screen
### Overview of change
- Created a common component, `VulnerabilityTodoList`, and deleted `ToDoMobile.jsx`.
- Updated the reference from `ToDoPage.jsx` to the new component.

### Reason for change
- Previously, different components were used for PC and mobile views, leading to duplicated code as their functionalities were very similar.
- By consolidating into a single component, we improve code maintainability and reusability, enabling more efficient future feature additions and UI enhancements.

### Technical details
- The mobile UI logic from `ToDoMobile.jsx` and the PC UI logic from `ToDoTable.jsx` were integrated and refactored into `VulnerabilityTodoList.jsx`.
- This new component uses `@mui/material`'s responsive design features to render an appropriate layout for both PC and mobile screen sizes within a single codebase.
- This results in a simplified codebase and a cleaner architecture.
